refactor(math): flatten directory structure, remove unused abstractions
- Flatten field module: `field::fields::fft_friendly::u64_goldilocks` → `field::goldilocks`,
  `field::fields::fft_friendly::extensions_goldilocks` → `field::extensions_goldilocks`
- Remove generic extension traits (CubicExtensionField, QuadraticExtensionField) —
  Goldilocks extensions implement IsField directly
- Remove UnsignedInteger<N>, U256, U384, and Montgomery backend — only used in own tests
- Move inline #[cfg(test)] blocks from source files to dedicated test files
- Clean up unused imports and dead code warnings

Deleted ~5,700 lines across 52 files. All 323 crypto tests pass.
