+++ crypto/crypto/src/merkle_tree/backends/field_element.rs
@@ -76,7 +76,8 @@ where
 mod tests {
     use alloc::vec::Vec;
     use math::field::{
-        element::FieldElement, fields::fft_friendly::u64_goldilocks::GoldilocksField,
+        element::FieldElement,
+        goldilocks::GoldilocksField,
     };
     use sha3::{Keccak256, Keccak512, Sha3_256, Sha3_512};
