+++ crypto/crypto/src/merkle_tree/backends/field_element_vector.rs
@@ -131,7 +131,8 @@ where
 #[cfg(test)]
 mod tests {
     use math::field::{
-        element::FieldElement, fields::fft_friendly::u64_goldilocks::GoldilocksField,
+        element::FieldElement,
+        goldilocks::GoldilocksField,
     };
     use sha2::Sha512;
     use sha3::{Keccak256, Keccak512, Sha3_256, Sha3_512};
