+++ crypto/crypto/src/tests/default_transcript_tests.rs
@@ -1,9 +1,8 @@
 use alloc::vec::Vec;
 use math::field::{
     element::FieldElement,
-    fields::fft_friendly::{
-        extensions_goldilocks::Degree3GoldilocksExtensionField, u64_goldilocks::GoldilocksField,
-    },
+    extensions_goldilocks::Degree3GoldilocksExtensionField,
+    goldilocks::GoldilocksField,
 };
 
 use crate::fiat_shamir::default_transcript::DefaultTranscript;
@@ -11,7 +10,8 @@ use crate::fiat_shamir::is_transcript::IsTranscript;
 
 #[test]
 fn basic_challenge() {
-    let mut transcript = DefaultTranscript::<Degree3GoldilocksExtensionField>::default();
+    let mut transcript =
+        DefaultTranscript::<Degree3GoldilocksExtensionField>::default();
 
     let point_a: Vec<u8> = vec![0xFF, 0xAB];
     let point_b: Vec<u8> = vec![0xDD, 0x8C, 0x9D];
@@ -53,27 +53,18 @@ type Ext3FE = FieldElement<Degree3GoldilocksExtensionField>;
 #[test]
 fn degree3_goldilocks_transcript_distinguish_different_fe() {
     let mut transcript_1 = DefaultTranscript::<Degree3GoldilocksExtensionField>::default();
-    transcript_1.append_field_element(&Ext3FE::new([
-        GoldFE::one(),
-        GoldFE::zero(),
-        GoldFE::zero(),
-    ]));
+    transcript_1
+        .append_field_element(&Ext3FE::new([GoldFE::one(), GoldFE::zero(), GoldFE::zero()]));
     let sample_1 = transcript_1.sample_field_element();
 
     let mut transcript_2 = DefaultTranscript::<Degree3GoldilocksExtensionField>::default();
-    transcript_2.append_field_element(&Ext3FE::new([
-        GoldFE::zero(),
-        GoldFE::zero(),
-        GoldFE::one(),
-    ]));
+    transcript_2
+        .append_field_element(&Ext3FE::new([GoldFE::zero(), GoldFE::zero(), GoldFE::one()]));
     let sample_2 = transcript_2.sample_field_element();
 
     let mut transcript_3 = DefaultTranscript::<Degree3GoldilocksExtensionField>::default();
-    transcript_3.append_field_element(&Ext3FE::new([
-        GoldFE::one(),
-        GoldFE::zero(),
-        GoldFE::zero(),
-    ]));
+    transcript_3
+        .append_field_element(&Ext3FE::new([GoldFE::one(), GoldFE::zero(), GoldFE::zero()]));
     let sample_3 = transcript_3.sample_field_element();
 
     assert!(sample_1 != sample_2);
@@ -83,7 +74,7 @@ fn degree3_goldilocks_transcript_distinguish_different_fe() {
 #[test]
 fn fork_determinism() {
     // Cloning a transcript twice and running the same operations must produce identical challenges.
-    let mut base = DefaultTranscript::<GoldilocksField>::default();
+    let mut base = DefaultTranscript::<FrField>::default();
     base.append_bytes(&[0x01, 0x02, 0x03]);
     let _ = base.sample();
     base.append_bytes(&[0xAA, 0xBB]);
@@ -101,7 +92,7 @@ fn fork_determinism() {
 #[test]
 fn fork_domain_separator_differentiates() {
     // Two forks from the same base with different domain separators must produce different challenges.
-    let mut base = DefaultTranscript::<GoldilocksField>::default();
+    let mut base = DefaultTranscript::<FrField>::default();
     base.append_bytes(&[0x01, 0x02, 0x03]);
     let _ = base.sample();
     base.append_bytes(&[0xAA, 0xBB]);
@@ -118,7 +109,7 @@ fn fork_domain_separator_differentiates() {
 #[test]
 fn fork_isolation() {
     // Appending data to one fork must not affect challenges sampled from another.
-    let mut base = DefaultTranscript::<GoldilocksField>::default();
+    let mut base = DefaultTranscript::<FrField>::default();
     base.append_bytes(&[0x01, 0x02, 0x03]);
     let _ = base.sample();
