--- crypto/math/src/field/extensions/mod.rs
@@ -1,2 +1,0 @@
-pub mod cubic;
-pub mod quadratic;
