--- crypto/math/src/field/fields/fft_friendly/mod.rs
@@ -1,4 +1,0 @@
-/// Quadratic and cubic extensions of the Goldilocks field
-pub mod extensions_goldilocks;
-/// Optimized Goldilocks field p = 2^64 - 2^32 + 1 (no Montgomery form)
-pub mod u64_goldilocks;
