--- crypto/math/src/field/fields/mod.rs
@@ -1,2 +1,0 @@
-/// Implementation of two-adic prime fields to use with the Fast Fourier Transform (FFT).
-pub mod fft_friendly;
