+++ crypto/math/src/tests/fft_tests.rs
@@ -54,7 +54,7 @@ mod fft_polynomial_tests {
     };
     use crate::fft::polynomial::compose_fft;
     use crate::field::element::FieldElement;
-    use crate::field::test_fields::u64_test_field::{U64TestField, U64TestFieldExtension};
+    use crate::field::extensions_goldilocks::Degree2GoldilocksExtensionField;
     use crate::field::traits::{IsFFTField, RootsConfig};
     use crate::polynomial::Polynomial;
     use proptest::{collection, prelude::*};
@@ -259,8 +259,9 @@ mod fft_polynomial_tests {
 
     #[test]
     fn test_fft_with_values_in_field_extension_over_domain_in_prime_field() {
-        type TF = U64TestField;
-        type TL = U64TestFieldExtension;
+        use crate::field::goldilocks::GoldilocksField;
+        type TF = GoldilocksField;
+        type TL = Degree2GoldilocksExtensionField;
 
         let a = FieldElement::<TL>::from(&[FieldElement::one(), FieldElement::one()]);
         let b = FieldElement::<TL>::from(&[-FieldElement::from(2), FieldElement::from(17)]);
