+++ crypto/math/src/tests/mod.rs
@@ -1,11 +1,6 @@
-pub mod extensions_cubic_tests;
-pub mod extensions_quadratic_tests;
 pub mod fft_friendly_extensions_goldilocks_tests;
 pub mod fft_friendly_u64_goldilocks_tests;
 pub mod fft_tests;
 pub mod field_element_tests;
 pub mod polynomial_tests;
 pub mod test_fields_tests;
-pub mod unsigned_integer_montgomery_tests;
-pub mod unsigned_integer_u256_tests;
-pub mod unsigned_integer_u384_tests;
