+++ crypto/math/src/unsigned_integer/mod.rs
@@ -1,6 +1,1 @@
-// By removing refs as clippy wants
-// Implementations with all the combination of reference and not references become recursive
-#[allow(clippy::op_ref)]
-pub mod element;
-pub mod montgomery;
 pub mod traits;
