+++ crypto/stark/src/constraints/boundary.rs
@@ -161,7 +161,9 @@ impl<F: IsField> BoundaryConstraints<F> {
 
 #[cfg(test)]
 mod test {
-    use math::field::{fields::fft_friendly::u64_goldilocks::GoldilocksField, traits::IsFFTField};
+    use math::field::{
+        goldilocks::GoldilocksField, traits::IsFFTField,
+    };
     type PrimeField = GoldilocksField;
 
     use super::*;
