+++ crypto/stark/src/examples/dummy_air.rs
@@ -11,7 +11,8 @@ use crate::{
     traits::{AIR, TransitionEvaluationContext},
 };
 use math::field::{
-    element::FieldElement, fields::fft_friendly::u64_goldilocks::GoldilocksField,
+    element::FieldElement,
+    goldilocks::GoldilocksField,
     traits::IsFFTField,
 };
