+++ crypto/stark/src/examples/fibonacci_2_cols_shifted.rs
@@ -256,7 +256,8 @@ pub fn compute_trace<F: IsFFTField>(
 #[cfg(test)]
 mod tests {
     use math::field::{
-        element::FieldElement, fields::fft_friendly::u64_goldilocks::GoldilocksField,
+        element::FieldElement,
+        goldilocks::GoldilocksField,
     };
 
     use super::compute_trace;
