+++ crypto/stark/src/prover.rs
@@ -1625,7 +1625,8 @@ mod tests {
     use super::*;
     use math::{
         field::{
-            element::FieldElement, fields::fft_friendly::u64_goldilocks::GoldilocksField,
+            element::FieldElement,
+            goldilocks::GoldilocksField,
             traits::IsFFTField,
         },
         polynomial::Polynomial,
@@ -1722,4 +1723,5 @@ mod tests {
             assert_eq!(*eval, poly.evaluate(&(offset * primitive_root.pow(i))));
         }
     }
+
 }
