+++ crypto/stark/src/tests/air_tests.rs
@@ -3,9 +3,7 @@
 use crypto::fiat_shamir::default_transcript::DefaultTranscript;
 use math::field::{
     element::FieldElement,
-    fields::fft_friendly::{
-        extensions_goldilocks::Degree3GoldilocksExtensionField, u64_goldilocks::GoldilocksField,
-    },
+    extensions_goldilocks::Degree3GoldilocksExtensionField, goldilocks::GoldilocksField,
 };
 
 use crate::traits::AIR;
