+++ crypto/stark/src/tests/bus_tests/packing_tests.rs
@@ -1,7 +1,7 @@
 //! Unit tests for Packing combine logic.
 
 use math::field::element::FieldElement;
-use math::field::fields::fft_friendly::u64_goldilocks::GoldilocksField;
+use math::field::goldilocks::GoldilocksField;
 
 use crate::lookup::{
     AirWithBuses, AuxiliaryTraceBuildData, BusInteraction, Multiplicity,
@@ -305,8 +305,7 @@ fn test_quad_wl_equals_four_direct() {
 
 #[test]
 fn test_air_layout_single_interaction() {
-    type E =
-        math::field::fields::fft_friendly::extensions_goldilocks::Degree3GoldilocksExtensionField;
+    type E = math::field::extensions_goldilocks::Degree3GoldilocksExtensionField;
 
     let interaction = BusInteraction::sender(
         TEST_BUS,
@@ -332,8 +331,7 @@ fn test_air_layout_single_interaction() {
 
 #[test]
 fn test_air_layout_multiple_interactions() {
-    type E =
-        math::field::fields::fft_friendly::extensions_goldilocks::Degree3GoldilocksExtensionField;
+    type E = math::field::extensions_goldilocks::Degree3GoldilocksExtensionField;
 
     let interaction1 = BusInteraction::sender(
         TEST_BUS,
